Create a chain for multiple sources
