+++ main.py
@@ -1,47 +1,37 @@
-import requests
 import api_keys
-from bs4 import BeautifulSoup
+from web_scraper_utils import web_search, scrape_text
 from langchain_openai import ChatOpenAI
 from langchain.prompts import ChatPromptTemplate
 from langchain.schema.output_parser import StrOutputParser
+from langchain.schema.runnable import RunnablePassthrough, RunnableLambda
 
-template = """
+template = """{text}
 
-Summarize the following question based on the context:
+--------------
 
-Question: {question}
+Using the above text, answer in short the following question:
 
-Context: {context}
+> {question}
 
+--------------
+if the question cannot be answered using the text, imply summarize the text. Include all factual information , numbers stats etc
 """
-
-
-def scrape_text(url: str):
-    # Get the text of a page
-    try:
-        response = requests.get(url)
-
-        if response.status_code == 200:
-            soup = BeautifulSoup(response.text, "html.parser")
-
-            page_text = soup.get_text(separator=" ", strip=True)
-
-            return page_text
-        else:
-            return f"Failed to retrieve the webpage: Status code {response.status_code}"
-    except Exception as e:
-        print(e)
 
 prompt = ChatPromptTemplate.from_template(template)
 
 url = "https://blog.langchain.dev/announcing-langsmith"
 pageContent = scrape_text(url)
-chain = prompt | ChatOpenAI(model="gpt-3.5-turbo-1106") | StrOutputParser()
+scrape_and_summarize_chain = RunnablePassthrough.assign(
+    text=lambda x: scrape_text(x["url"])[:10000]
+) | prompt | ChatOpenAI(model="gpt-3.5-turbo-1106") | StrOutputParser()
+
+chain = RunnablePassthrough.assign(
+    urls=lambda x: web_search(x["question"])
+) | (lambda x: [{"question": x["question"], "url": u} for u in x["urls"]]) | scrape_and_summarize_chain.map()
 
 response = chain.invoke(
     {
-        "question": "what is langsmith",
-        "context": pageContent
+        "question": "what is langsmith"
     }
 )
